remove V1 directory
