+++ app/controllers/api/alcohols_controller.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Api
+  class AlcoholsController < ApplicationController
+    def index; end
+  end
+end
