+++ app/controllers/api/diagnoses_controller.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Api
+  class DiagnosesController < ApplicationController
+    def index; end
+
+    def create; end
+  end
+end
