--- app/controllers/api/v1/alcohols_controller.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Api
-  module V1
-    class AlcoholsController < ApplicationController
-      def index; end
-    end
-  end
-end
