--- app/controllers/api/v1/diagnoses_controller.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Api
-  module V1
-    class DiagnosesController < ApplicationController
-      def index; end
-
-      def create; end
-    end
-  end
-end
