--- app/controllers/api/v1/users_controller.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-module Api
-  module V1
-    class UsersController < ApplicationController
-      def index
-        @users = User.all
-        render json: @users
-      end
-    end
-  end
-end
