+++ config/routes.rb
@@ -4,9 +4,7 @@ Rails.application.routes.draw do
   devise_for :users
 
   namespace :api do
-    namespace :v1 do
-      resources :users, only: %i[index]
-    end
+    resources :users, only: %i[index]
   end
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
 end
