+++ spec/requests/api/diagnoses_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe 'Api::V1::Alcohols', type: :request do
+RSpec.describe 'Api::Diagnoses', type: :request do
   describe 'GET /index' do
     pending "add some examples (or delete) #{__FILE__}"
   end
