+++ spec/requests/api/users_spec.rb
@@ -2,10 +2,10 @@
 
 require 'rails_helper'
 
-RSpec.describe 'Api::V1::Users', type: :request do
-  describe 'GET /api/v1/users' do
+RSpec.describe 'Api::Users', type: :request do
+  describe 'GET /api/users' do
     it 'works! (now write some real specs)' do
-      get api_v1_users_path
+      get api_users_path
       expect(response).to have_http_status(200)
     end
   end
